tests/select/ refactor
