+++ tests/constants.py
@@ -1,0 +1,4 @@
+import pandas as pd
+
+
+NWP_FREQ = pd.Timedelta("3h")
